Database entegrasyonu ve WebUI düzenlemesi

+++ Entities/Concrete/User.cs
@@ -13,7 +13,7 @@ namespace Entities.Concrete
         public string LastName { get; set; }
         public string Email { get; set; }
         public string Password { get; set; }
-        public BigInteger Phone { get; set; }
+        public UInt64 Phone { get; set; }
         public int Role { get; set; }
     }
 }
